feat: #6014 small refactor
