+++ packages/developer-portal/src/components/pages/apps/new/__tests__/auth-options-content.test.tsx
