+++ packages/developer-portal/src/components/pages/apps/new/__tests__/permissions-options-content.test.tsx
