+++ packages/developer-portal/src/components/pages/apps/new/apps-new.tsx
@@ -21,15 +21,15 @@ import { yupResolver } from '@hookform/resolvers/yup'
 import { useForm, UseFormTrigger } from 'react-hook-form'
 import { object, SchemaOf, string, TestFunction } from 'yup'
 import { isValidUrlWithCustomScheme, UpdateActionNames, updateActions } from '@reapit/utils-common'
-import errorMessages from '../../../constants/error-messages'
+import errorMessages from '../../../../constants/error-messages'
 import { AnyObject } from 'yup/lib/types'
 import { AppDetailModel, CreateAppModel } from '@reapit/foundations-ts-definitions'
 import generate from 'project-name-generator'
 import { ReapitConnectSession, useReapitConnect } from '@reapit/connect-session'
 import dashify from 'dashify'
-import { reapitConnectBrowserSession } from '../../../core/connect-session'
+import { reapitConnectBrowserSession } from '../../../../core/connect-session'
 import { useReapitUpdate, SendFunction } from '@reapit/utils-react'
-import Routes from '../../../constants/routes'
+import Routes from '../../../../constants/routes'
 import { History } from 'history'
 import { useHistory } from 'react-router-dom'
 import { HelperContent } from './helper-content'
