+++ packages/developer-portal/src/components/pages/apps/new/permissions-options-content.tsx
@@ -7,7 +7,7 @@ import { useReapitConnect } from '@reapit/connect-session'
 import { ScopeModel } from '@reapit/foundations-ts-definitions'
 import { getActions, GetActionNames } from '@reapit/utils-common'
 import { useReapitGet } from '@reapit/utils-react'
-import { reapitConnectBrowserSession } from '../../../core/connect-session'
+import { reapitConnectBrowserSession } from '../../../../core/connect-session'
 import { UseFormRegister, UseFormGetValues } from 'react-hook-form'
 import { CreateAppFormSchema } from './apps-new'
 import debounce from 'just-debounce-it'
